requirements.txt
numpy
onnxruntime-gpu
soundfile
librosa
tokenizers
fastapi
uvicorn[standard]
websockets